Internationalization
Trying to implement a language change option

+++ ui/lib/src/pages/settings/settings_model.dart
@@ -5,7 +5,7 @@ class Settings {
   // Audio Settings
   String inputDevice;
   String outputDevice;
-  int inputSensitivity; // Start with 5, modify later
+  double inputSensitivity;
 
   // 3 Param Constructor
   Settings(this.inputDevice, this.outputDevice, this.inputSensitivity);
